Address parser compatibility review feedback
